Fix DetailBook
Fix Grade return type,
double -> Double Object Type

+++ src/main/java/com/ryan/domain/BookGradeVO.java
@@ -14,5 +14,5 @@ public class BookGradeVO {
     private int bookNum;
 
     // 평점 
-    private int gradeScore;
+    private double gradeScore;
 }
